+++ lib/edfize/edf.rb
@@ -58,6 +58,13 @@ module Edfize
       get_data_records
     end
 
+    # Epic Number is Zero Indexed, and Epic Size is in Seconds (Not Data Records)
+    def load_epic(epic_number, epic_size)
+      # reset_signals!
+      load_digital_signals_by_epic(epic_number, epic_size)
+      calculate_physical_values!
+    end
+
     def size_of_header
       HEADER_OFFSET + ns * Signal::SIGNAL_CONFIG.collect{|k,h| h[:size]}.inject(:+)
     end
@@ -153,6 +160,11 @@ module Edfize
       @number_of_signals
     end
 
+    def reset_signals!
+      @signals = []
+      read_signal_header
+    end
+
     def create_signals
       (0..ns-1).to_a.each do |signal_number|
         @signals[signal_number] ||= Signal.new()
@@ -190,12 +202,25 @@ module Edfize
       calculate_physical_values!()
     end
 
+    def load_digital_signals_by_epic(epic_number, epic_size)
+      size_of_data_record_in_bytes = @signals.collect(&:samples_per_data_record).inject(:+).to_i * SIZE_OF_SAMPLE_IN_BYTES
+      data_records_to_retrieve = (epic_size / @duration_of_a_data_record rescue 0)
+      length_of_bytes_to_read = data_records_to_retrieve * size_of_data_record_in_bytes
+      epic_offset_size = epic_number * size_of_data_record_in_bytes # TODO: The size in bytes of an epic
+
+      all_signal_data = IO.binread(@filename, length_of_bytes_to_read, size_of_header + epic_offset_size).unpack('s<*')
+      load_signal_data(all_signal_data, data_records_to_retrieve)
+    end
+
     # 16-bit signed integer size = 2 Bytes = 2 ASCII characters
     # 16-bit signed integer in "Little Endian" format (least significant byte first)
     # unpack:  s<         16-bit signed, (little-endian) byte order
     def load_digital_signals
-      @all_signal_data = IO.binread(@filename, nil, size_of_header).unpack('s<*')
+      all_signal_data = IO.binread(@filename, nil, size_of_header).unpack('s<*')
+      load_signal_data(all_signal_data, @number_of_data_records)
+    end
 
+    def load_signal_data(all_signal_data, data_records_retrieved)
       all_samples_per_data_record = @signals.collect{|s| s.samples_per_data_record}
       total_samples_per_data_record = all_samples_per_data_record.inject(:+).to_i
 
@@ -206,11 +231,11 @@ module Edfize
         offset += samples_per_data_record
       end
 
-      (0..@number_of_data_records-1).to_a.each do |data_record_index|
+      (0..data_records_retrieved-1).to_a.each do |data_record_index|
         @signals.each_with_index do |signal, signal_index|
           read_start = data_record_index * total_samples_per_data_record + offsets[signal_index]
           (0..signal.samples_per_data_record - 1).to_a.each do |value_index|
-            signal.digital_values << @all_signal_data[read_start+value_index]
+            signal.digital_values << all_signal_data[read_start+value_index]
           end
         end
       end