+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.2.0
 
+### Enhancements
+- EDFs can now be partially loaded by specifying the epic number and epic size in seconds
+
 ### Testing
 - Updated to minitest test framework along with simplecov for coverage